+++ frontend/src/components/CreateSection/CreateSection.js
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import CreateTask from "../CreateTask/CreateTask";
+import "./CreateSection.scss";
 
 const CreateSection = (props) => {
   const {
