+++ frontend/src/components/CreateSection/CreateSection.scss
@@ -1,0 +1,42 @@
+.CreateSection {
+  position: relative;
+  left: 15%;
+  width: 77%;
+  border: 1px solid rgb(87, 82, 82);
+  box-sizing: border-box;
+  padding: 5px;
+  margin: 5px 5px 5px 5px;
+
+  .up-section-button,
+  .down-section-button {
+    position: relative;
+    height: 30px;
+    left: -3%;
+    margin: 5px 5px 5px 5px;
+  }
+
+  .create-task-button {
+    position: relative;
+    height: 30px;
+    right: -10%;
+  }
+
+  .delete-section-button {
+    position: relative;
+    height: 30px;
+    right: 40%;
+    margin: 5px 5px 5px 5px;
+  }
+
+  .section-title {
+    position: relative;
+    width: 60%;
+    border: 1px solid rgb(87, 82, 82);
+    box-sizing: border-box;
+    padding: 5px;
+    margin: 15px 5px 15px 5px;
+    font-weight: 700;
+    color: rgb(0, 0, 225);
+    font-size: 20px;
+  }
+}
