+++ frontend/src/components/CreateTask/CreateTask.scss
@@ -1,0 +1,54 @@
+.CreateTask {
+  position: relative;
+  left: 20%;
+  width: 77%;
+  border: 1px solid rgb(87, 82, 82);
+  box-sizing: border-box;
+  padding: 5px;
+  margin: 5px 5px 5px 5px;
+
+  .task-type {
+    position: relative;
+    border: 1px solid rgb(87, 82, 82);
+    box-sizing: border-box;
+    padding: 5px;
+    margin: 5px 5px 5px 5px;
+    font-weight: 700;
+    color: rgb(0, 0, 225);
+    font-size: 20px;
+  }
+  .task-title,
+  .task-url,
+  .task-description {
+    position: relative;
+    width: 60%;
+    border: 1px solid rgb(87, 82, 82);
+    box-sizing: border-box;
+    padding: 5px;
+    margin: 5px 5px 5px 5px;
+    font-weight: 700;
+    color: rgb(0, 0, 225);
+    font-size: 20px;
+  }
+  label {
+    position: relative;
+    margin: 10px;
+    color: black;
+    font-size: 15px;
+  }
+
+  .up-task-button,
+  .down-task-button {
+    position: relative;
+    height: 30px;
+    left: -40%;
+    margin: 5px 5px 5px 5px;
+  }
+
+  .delete-task-button {
+    position: relative;
+    height: 30px;
+    right: -40%;
+    margin: 5px 5px 5px 5px;
+  }
+}
