+++ frontend/src/containers/CreateRoadmap/CreateRoadmap.js
@@ -1,301 +1,20 @@
 import React, { Component } from "react";
-// import FormGroup from "@material-ui/core/FormGroup";
-// import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Switch from "@material-ui/core/Switch";
 import { connect } from "react-redux";
 import { withRouter } from "react-router";
 import PropTypes from "prop-types";
 import * as actionCreators from "../../store/actions/index";
 
-import CreateSection from "../../components/CreateSection/CreateSection";
-import { levelType } from "../../constants";
-import "./CreateRoadmap.scss";
+import Roadmap from "../Roadmap/Roadmap";
 
 class CreateRoadmap extends Component {
-  state = {
-    isPrivate: false,
-    title: "",
-    level: 0,
-    description: "",
-    sections: [],
-    tags: [],
-    newTag: "",
-  };
-
-  onClickPrivate = () => {
-    const { isPrivate } = this.state;
-    this.setState({ isPrivate: !isPrivate });
-  };
-
-  onChangeTitle = (title) => {
-    this.setState({ title });
-  };
-
-  onChangeLevel = (level) => {
-    this.setState({ level: parseInt(level, 10) });
-  };
-
-  onChangeDescription = (description) => {
-    this.setState({ description });
-  };
-
-  onChangeNewTag = (newTag) => {
-    this.setState({ newTag });
-  };
-
-  onClickAddTag = () => {
-    const { tags, newTag } = this.state;
-    this.setState({ tags: tags.concat(newTag) });
-    this.setState({ newTag: "" });
-  };
-
-  onClickDeleteTag = (id) => {
-    const { tags } = this.state;
-    this.setState({ tags: tags.filter((_, index) => index !== id) });
-  };
-
-  onClickCreateSection = () => {
-    const { sections } = this.state;
-    this.setState({ sections: sections.concat({ section_title: "", tasks: [] }) });
-  };
-
-  onClickDeleteSection = (tmpSectionId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.filter((_, index) => {
-        return index !== tmpSectionId;
-      }),
-    });
-  };
-
-  onClickUpSection = (tmpSectionId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, index) => {
-        if (index === tmpSectionId - 1) {
-          return sections[index + 1];
-        }
-        if (index === tmpSectionId) {
-          return sections[index - 1];
-        }
-        return section;
-      }),
-    });
-  };
-
-  onClickDownSection = (tmpSectionId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, index) => {
-        if (index === tmpSectionId) {
-          return sections[index + 1];
-        }
-        if (index === tmpSectionId + 1) {
-          return sections[index - 1];
-        }
-        return section;
-      }),
-    });
-  };
-
-  onChangeSectionTitle = (tmpSectionId, title) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, index) => {
-        if (index === tmpSectionId) {
-          section.section_title = title;
-        }
-        return section;
-      }),
-    });
-  };
-
-  onClickCreateTask = (tmpSectionId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, index) => {
-        if (index === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.concat({
-              task_title: "",
-              task_type: 0,
-              task_url: "",
-              task_description: "",
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onClickDeleteTask = (tmpSectionId, tmpTaskId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.filter((_, taskIndex) => {
-              return taskIndex !== tmpTaskId;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onClickUpTask = (tmpSectionId, tmpTaskId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId - 1) {
-                return section.tasks[taskIndex + 1];
-              }
-              if (taskIndex === tmpTaskId) {
-                return section.tasks[taskIndex - 1];
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onClickDownTask = (tmpSectionId, tmpTaskId) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId) {
-                return section.tasks[taskIndex + 1];
-              }
-              if (taskIndex === tmpTaskId + 1) {
-                return section.tasks[taskIndex - 1];
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onChangeTaskTitle = (tmpSectionId, tmpTaskId, title) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId) {
-                return { ...task, task_title: title };
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onChangeTaskType = (tmpSectionId, tmpTaskId, type) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId) {
-                return { ...task, task_type: parseInt(type, 10) };
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onChangeTaskUrl = (tmpSectionId, tmpTaskId, url) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId) {
-                return { ...task, task_url: url };
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
-  onChangeTaskDescription = (tmpSectionId, tmpTaskId, description) => {
-    const { sections } = this.state;
-    this.setState({
-      sections: sections.map((section, sectionIndex) => {
-        if (sectionIndex === tmpSectionId) {
-          return {
-            ...section,
-            tasks: section.tasks.map((task, taskIndex) => {
-              if (taskIndex === tmpTaskId) {
-                return { ...task, task_description: description };
-              }
-              return task;
-            }),
-          };
-        }
-        return section;
-      }),
-    });
-  };
-
   onClickCreateBack = () => {
     const { history } = this.props;
-    const { title, level, description, sections } = this.state;
-    if (title !== "" || level !== 0 || sections.length !== 0 || description !== "") {
-      const back = window.confirm("Leave the page? Changes you made will be deleted.");
-      if (back) {
-        history.goBack();
-      }
-    } else history.goBack();
+    history.goBack();
   };
 
-  onClickCreateConfirm = () => {
-    const { isPrivate, title, level, description, sections, tags } = this.state;
+  onClickCreateConfirm = (roadmapData) => {
     const { onCreateRoadmap } = this.props;
-    const roadmapData = {
-      private: isPrivate,
-      title,
-      level,
-      description,
-      sections,
-      tags,
-    };
-    onCreateRoadmap(roadmapData);
+    onCreateRoadmap({ ...roadmapData, addedTagList: undefined, deletedTagList: undefined });
   };
 
   render() {
@@ -306,131 +25,28 @@ class CreateRoadmap extends Component {
       return <div />;
     }
 
-    const { isPrivate, sections, level, title, description, tags, newTag } = this.state;
-
-    const taglist = tags.map((tag, index) => {
-      return (
-        <div className="tags">
-          <p key={tag}>{tag}</p>
-          <button className="delete-tag" type="button" onClick={() => this.onClickDeleteTag(index)}>
-            delete
-          </button>
-        </div>
-      );
-    });
-
-    const CreateSections = sections.map((section, index) => {
-      return (
-        <CreateSection
-          tmpSectionId={index}
-          sectionLastId={sections.length - 1}
-          title={section.section_title}
-          tasks={section.tasks}
-          clickDeleteSectionHandler={this.onClickDeleteSection}
-          clickUpSectionHandler={this.onClickUpSection}
-          clickDownSectionHandler={this.onClickDownSection}
-          changeSectionTitleHandler={this.onChangeSectionTitle}
-          clickCreateTaskHandler={this.onClickCreateTask}
-          clickDeleteTaskHandler={this.onClickDeleteTask}
-          clickUpTaskHandler={this.onClickUpTask}
-          clickDownTaskHandler={this.onClickDownTask}
-          changeTaskTitleHandler={this.onChangeTaskTitle}
-          changeTaskTypeHandler={this.onChangeTaskType}
-          changeTaskUrlHandler={this.onChangeTaskUrl}
-          changeTaskDescriptionHandler={this.onChangeTaskDescription}
-        />
-      );
-    });
-
     return (
       <div className="CreateRoadmap">
-        <h1 className="header">Create Roadmap</h1>
-        <div className="roadmap">
-          <label>
-            Private
-            <Switch
-              id="roadmap-private"
-              checked={isPrivate}
-              onClick={() => this.onClickPrivate()}
-            />
-          </label>
-          <label>Title</label>
-          <input
-            id="roadmap-title"
-            type="text"
-            value={title}
-            onChange={(event) => this.onChangeTitle(event.target.value)}
-          />
-          <select
-            id="roadmap-level"
-            value={level}
-            onChange={(event) => {
-              return this.onChangeLevel(event.target.value);
-            }}
-          >
-            Level
-            <option value={0}>Choose level</option>
-            <option value={levelType.BASIC}>Basic</option>
-            <option value={levelType.INTERMEDIATE}>Intermediate</option>
-            <option value={levelType.ADVANCED}>Advanced</option>
-          </select>
-          <br />
-          <label>Tags</label>
-          {taglist}
-          <input
-            id="new-tag"
-            value={newTag}
-            onChange={(event) => this.onChangeNewTag(event.target.value)}
-          />
-          <button id="add-tag-button" type="button" onClick={() => this.onClickAddTag()}>
-            add
-          </button>
-          <br />
-          <label>Description</label>
-          <input
-            id="roadmap-description"
-            value={description}
-            onChange={(event) => this.onChangeDescription(event.target.value)}
-          />
-        </div>
-        <div className="sections">
-          {CreateSections}
-          <button
-            type="button"
-            id="create-section-button"
-            onClick={() => {
-              this.onClickCreateSection();
-            }}
-          >
-            Create Section
-          </button>
-        </div>
-        <div className="buttons">
-          <button
-            id="back-create-roadmap-button"
-            type="button"
-            onClick={() => this.onClickCreateBack()}
-          >
-            Back
-          </button>
-          <button
-            id="confirm-create-roadmap-button"
-            type="button"
-            disabled={title === "" || level === 0 || sections.length === 0}
-            onClick={() => this.onClickCreateConfirm()}
-          >
-            Confirm
-          </button>
-        </div>
+        <Roadmap
+          isEdit={false}
+          onClickBackHandler={this.onClickCreateBack}
+          onClickConfirmHandler={this.onClickCreateConfirm}
+        />
       </div>
     );
   }
 }
 
 CreateRoadmap.propTypes = {
-  history: PropTypes.objectOf(PropTypes.any),
   selectedUser: PropTypes.objectOf(PropTypes.any),
   onCreateRoadmap: PropTypes.func.isRequired,
+  history: PropTypes.objectOf(PropTypes.any),
+};
+
+const mapStateToProps = (state) => {
+  return {
+    selectedUser: state.user.selectedUser,
+  };
 };
 
 const mapDispatchToProps = (dispatch) => {
@@ -439,4 +55,4 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(null, mapDispatchToProps)(withRouter(CreateRoadmap));
+export default connect(mapStateToProps, mapDispatchToProps)(withRouter(CreateRoadmap));
