+++ frontend/src/containers/EditRoadmap/EditRoadmap.test.js
@@ -6,6 +6,7 @@ import { Route, Switch } from "react-router-dom";
 import { ConnectedRouter } from "connected-react-router";
 import getMockStore from "../../test-utils/mocks";
 import { history } from "../../store/store";
+import Roadmap from "../Roadmap/Roadmap";
 import EditRoadmap from "./EditRoadmap";
 import * as actionCreators from "../../store/actions/roadmap";
 
@@ -149,7 +150,7 @@ describe("<EditRoadmap />", () => {
       </Provider>
     );
     const component = mount(tmpEditRoadmap);
-    const wrapper = component.find(".CreateRoadmap");
+    const wrapper = component.find(".EditRoadmap");
     expect(wrapper.length).toBe(0);
     expect(spyAlert).toHaveBeenCalledTimes(1);
     spyAlert.mockRestore();
@@ -201,7 +202,7 @@ describe("<EditRoadmap />", () => {
 
   it("should set state if selectedRoadmap has been received", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state).toEqual({
       ...initialRoadmapState.selectedRoadmap,
       received: true,
@@ -231,7 +232,7 @@ describe("<EditRoadmap />", () => {
     wrapper.at(2).simulate("click");
     wrapper = component.find("#roadmap-description");
     wrapper.simulate("change", { target: { value: description } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.title).toBe(title);
     expect(instance.state.level).toBe(level);
     expect(instance.state.newTag).toBe(newTag);
@@ -246,7 +247,7 @@ describe("<EditRoadmap />", () => {
     wrapper.simulate("change", { target: { value: newTag } });
     wrapper = component.find("#add-tag-button");
     wrapper.simulate("click");
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.tags).toEqual(["tag0", "tag1", newTag]);
     expect(instance.state.addedTagList).toEqual([newTag]);
     wrapper = component.find(".delete-tag");
@@ -257,7 +258,7 @@ describe("<EditRoadmap />", () => {
 
   it("should call 'onClickCreateSection'", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     const sectionsNum = instance.state.sections.length;
     const wrapper2 = component.find("#create-section-button");
     wrapper2.simulate("click");
@@ -266,7 +267,7 @@ describe("<EditRoadmap />", () => {
 
   it("should call 'onClickDeleteSection'", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     const sectionsNum = instance.state.sections.length;
     const wrapper = component.find(".delete-section-button");
     wrapper.at(0).simulate("click");
@@ -277,7 +278,7 @@ describe("<EditRoadmap />", () => {
     const component = mount(editRoadmap);
     const wrapper = component.find(".up-section-button");
     wrapper.at(1).simulate("click");
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].section_title).toBe("test-section1");
     expect(instance.state.sections[1].section_title).toBe("test-section0");
     expect(instance.state.sections[2].section_title).toBe("test-section2");
@@ -287,7 +288,7 @@ describe("<EditRoadmap />", () => {
     const component = mount(editRoadmap);
     const wrapper = component.find(".down-section-button");
     wrapper.at(0).simulate("click");
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].section_title).toBe("test-section1");
     expect(instance.state.sections[1].section_title).toBe("test-section0");
     expect(instance.state.sections[2].section_title).toBe("test-section2");
@@ -297,7 +298,7 @@ describe("<EditRoadmap />", () => {
     const component = mount(editRoadmap);
     const wrapper = component.find(".section-title");
     wrapper.at(0).simulate("change", { target: { value: "55" } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].section_title).toBe("55");
     expect(instance.state.sections[1].section_title).toBe("test-section1");
   });
@@ -306,7 +307,7 @@ describe("<EditRoadmap />", () => {
     const component = mount(editRoadmap);
     const wrapper = component.find(".create-task-button");
     wrapper.at(0).simulate("click");
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].tasks[3]).toEqual({
       task_title: "",
       task_type: 0,
@@ -318,7 +319,7 @@ describe("<EditRoadmap />", () => {
 
   it("should call 'onClickDeleteTask'", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     const tasksNum0 = instance.state.sections[0].tasks.length;
     const tasksNum1 = instance.state.sections[1].tasks.length;
     const wrapper = component.find(".delete-task-button");
@@ -332,7 +333,7 @@ describe("<EditRoadmap />", () => {
     const testTitle = "test";
     const wrapper = component.find(".task-title");
     wrapper.at(0).simulate("change", { target: { value: testTitle } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].tasks[0].task_title).toEqual(testTitle);
     expect(instance.state.sections[0].tasks[1].task_title).toEqual("task1-title");
     expect(instance.state.sections[1]).toEqual({
@@ -359,7 +360,7 @@ describe("<EditRoadmap />", () => {
     const testType = 2;
     const wrapper = component.find(".task-type");
     wrapper.at(0).simulate("change", { target: { value: testType } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].tasks[0].task_type).toEqual(testType);
     expect(instance.state.sections[0].tasks[1].task_type).toEqual(2);
     expect(instance.state.sections[1]).toEqual({
@@ -386,7 +387,7 @@ describe("<EditRoadmap />", () => {
     const testUrl = "test";
     const wrapper = component.find(".task-url");
     wrapper.at(0).simulate("change", { target: { value: testUrl } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].tasks[0].task_url).toEqual(testUrl);
     expect(instance.state.sections[0].tasks[1].task_url).toEqual("task1-url");
     expect(instance.state.sections[1]).toEqual({
@@ -413,7 +414,7 @@ describe("<EditRoadmap />", () => {
     const testDescription = "test";
     const wrapper = component.find(".task-description");
     wrapper.at(0).simulate("change", { target: { value: testDescription } });
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     expect(instance.state.sections[0].tasks[0].task_description).toEqual(testDescription);
     expect(instance.state.sections[0].tasks[1].task_description).toEqual("task1-des");
     expect(instance.state.sections[1]).toEqual({
@@ -437,7 +438,7 @@ describe("<EditRoadmap />", () => {
 
   it("should call 'onClickUpTask'", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     const wrapper = component.find(".up-task-button");
     wrapper.at(1).simulate("click");
     expect(instance.state.sections[0].tasks[0].task_title).toBe("task1-title");
@@ -464,7 +465,7 @@ describe("<EditRoadmap />", () => {
 
   it("should call 'onClickDownTask'", () => {
     const component = mount(editRoadmap);
-    const instance = component.find(EditRoadmap.WrappedComponent).instance();
+    const instance = component.find(Roadmap).instance();
     const wrapper = component.find(".down-task-button");
     wrapper.at(0).simulate("click");
     expect(instance.state.sections[0].tasks[0].task_title).toBe("task1-title");
@@ -492,7 +493,7 @@ describe("<EditRoadmap />", () => {
   it("should call 'onClickCreateBack' - confirm=True", () => {
     const spyBackConfirmTrue = jest.spyOn(window, "confirm").mockImplementation(() => true);
     const component = mount(editRoadmap);
-    const wrapper = component.find("#back-edit-roadmap-button");
+    const wrapper = component.find("#back-roadmap-button");
     wrapper.simulate("click");
     expect(spyBackConfirmTrue).toHaveBeenCalledTimes(1);
     expect(spyHistoryGoBack).toHaveBeenCalledTimes(1);
@@ -502,7 +503,7 @@ describe("<EditRoadmap />", () => {
   it("should call 'onClickEditBack' - confirm=False", () => {
     const spyBackConfirmFalse = jest.spyOn(window, "confirm").mockImplementation(() => false);
     const component = mount(editRoadmap);
-    const wrapper = component.find("#back-edit-roadmap-button");
+    const wrapper = component.find("#back-roadmap-button");
     wrapper.simulate("click");
     expect(spyBackConfirmFalse).toHaveBeenCalledTimes(1);
     expect(spyHistoryGoBack).toHaveBeenCalledTimes(0);
@@ -512,7 +513,7 @@ describe("<EditRoadmap />", () => {
   it("should call 'onClickEditConfirm'", () => {
     const spyEditRoadmap = jest.spyOn(actionCreators, "editRoadmap");
     const component = mount(editRoadmap);
-    const wrapper = component.find("#confirm-edit-roadmap-button");
+    const wrapper = component.find("#confirm-roadmap-button");
     wrapper.simulate("click");
     expect(spyEditRoadmap).toHaveBeenCalledTimes(1);
   });
