+++ frontend/src/store/actions/user.test.js
@@ -73,7 +73,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signIn()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
@@ -92,7 +91,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signIn()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
@@ -126,7 +124,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signOut()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
@@ -145,7 +142,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signOut()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
@@ -179,7 +175,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signUp()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
@@ -198,7 +193,6 @@ describe("ActionCreatorsUser", () => {
     store.dispatch(actionCreators.signUp()).then(() => {
       expect(spy).toHaveBeenCalledTimes(1);
       expect(spyAlert).toHaveBeenCalledTimes(1);
-      spyAlert.mockRestore();
       done();
     });
   });
